requirements.txt
asgiref==3.6.0
astroid==2.15.5
autopep8==2.0.2
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==41.0.1
defusedxml==0.7.1
dill==0.3.6
Django==4.2.1
django-cors-headers==4.0.0
django-debug-toolbar==4.0.0
django-filter==23.2
django-khalti==1.0
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
djoser==2.2.0
drf-nested-routers==0.93.4
idna==3.4
isort==5.12.0
lazy-object-proxy==1.9.0
mccabe==0.7.0
mysqlclient==2.0.3
oauthlib==3.2.2
pep8==1.7.1
Pillow==9.4.0
pip==23.0.1
platformdirs==3.5.1
pycodestyle==2.10.0
pycparser==2.21
PyJWT==2.7.0
pylint==2.17.4
pylint-django==2.5.3
pylint-plugin-utils==0.8.1
python3-openid==3.2.0
pytz==2023.3
requests==2.31.0
requests-oauthlib==1.3.1
setuptools==66.0.0
social-auth-app-django==5.2.0
social-auth-core==4.4.2
sqlparse==0.4.4
tomlkit==0.11.8
typing_extensions==4.5.0
urllib3==2.0.2
wheel==0.38.4
wrapt==1.15.0
